gitignore et ajout des sauces

+++ backend/controllers/sauces.js
@@ -1,34 +1,81 @@
-let Sauces = require("../models/sauces");
+const Sauces = require("../models/sauces");
+const fs = require("fs");
 
 exports.createSauce = (req, res, next) => {
-  delete req.body._id;
-  const sauce = new Sauce({
-    ...req.body,
+  const sauceObject = JSON.parse(req.body.sauce);
+  delete sauceObject._id;
+  delete sauceObject._userId;
+  const sauce = new Sauces({
+    ...sauceObject,
+    userId: req.auth.userId,
+    imageUrl: `${req.protocol}://${req.get("host")}/images/${
+      req.file.filename
+    }`,
   });
+
   sauce
     .save()
-    .then(() => res.status(201).json({message: "Objet enregistré !"}))
-    .catch((error) => res.status(400).json({error}));
-  Sauces.updateOne({_id: req.params.id}, {...req.body, _id: req.params.id})
-    .then(() => res.status(200).json({message: "Objet modifié !"}))
-    .catch((error) => res.status(400).json({error}));
+    .then(() => {
+      res.status(201).json({message: "Votre sauce a bien été enregistrée !"});
+    })
+    .catch((error) => {
+      res.status(400).json({error});
+    });
 };
 
 exports.modifySauce = (req, res, next) => {
-  Sauces.updateOne({_id: req.params.id}, {...req.body, _id: req.params.id})
-    .then(() => res.status(200).json({message: "Votre sauce a été modifiée !"}))
-    .catch((error) => res.status(400).json({error}));
+  const sauceObject = req.file
+    ? {
+        ...JSON.parse(req.body.sauce),
+        imageUrl: `${req.protocol}://${req.get("host")}/images/${
+          req.file.filename
+        }`,
+      }
+    : {...req.body};
+
+  delete sauceObject._userId;
+  Sauces.findOne({_id: req.params.id})
+    .then((sauce) => {
+      if (sauce.userId != req.auth.userId) {
+        res.status(401).json({message: "Not authorized"});
+      } else {
+        Sauces.updateOne(
+          {_id: req.params.id},
+          {...sauceObject, _id: req.params.id}
+        )
+          .then(() => res.status(200).json({message: "Objet modifié!"}))
+          .catch((error) => res.status(401).json({error}));
+      }
+    })
+    .catch((error) => {
+      res.status(400).json({error});
+    });
 };
 
 exports.deleteSauce = (req, res, next) => {
-  Sauces.deleteOne({_id: req.params.id})
-    .then(() => res.status(200).json({message: "Votre sauce supprimée !"}))
-    .catch((error) => res.status(400).json({error}));
+  Sauces.findOne({_id: req.params.id})
+    .then((sauce) => {
+      if (sauce.userId != req.auth.userId) {
+        res.status(401).json({message: "Suppression non autorisée"});
+      } else {
+        const filename = sauce.imageUrl.split("/images/")[1];
+        fs.unlink(`images/${filename}`, () => {
+          Sauces.deleteOne({_id: req.params.id})
+            .then(() => {
+              res.status(200).json({message: "Objet supprimé !"});
+            })
+            .catch((error) => res.status(401).json({error}));
+        });
+      }
+    })
+    .catch((error) => {
+      res.status(500).json({error});
+    });
 };
 
 exports.getOneSauce = (req, res, next) => {
   Sauces.findOne({_id: req.params.id})
-    .then((sauce) => res.status(200).json(thing))
+    .then((sauce) => res.status(200).json(sauce))
     .catch((error) => res.status(404).json({error}));
 };
